# src 2/main/resources/openapi/app-api.yaml
openapi: 3.0.3
info:
  title: Order IT App API
  description: |-
    API for Order IT Application
  contact:
    email: info@mycloudpa.com
  version: 1.0.0
...
paths:
  /api/files/upload/company-logos:
    post:
      tags:
        - fileSystem
      summary: upload company logo
      description: upload company logo
      operationId: companyLogoUpload
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary

      responses:
        '201':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Invalid status value
  /api/files/upload/product-images:
    post:
      tags:
        - fileSystem
      summary: upload product image
      description: upload product image
      operationId: productImageUpload
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                oldUrl:
                  type: string
                  format: string
      responses:
        '201':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Invalid status value
...
  /api/favouriteSuppliers:
    post:
      tags:
        - favouriteSupplier
      summary: create a new Favourite Supplier
      description: create a new Favourite Supplier
      operationId: createFavouriteSupplier
      requestBody:
        description: new favourite supplier to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavouriteSupplier'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouriteSupplier'
        '400':
          description: Invalid status value
    put:
      tags:
        - favouriteSupplier
      summary: update an existing Favourite Supplier
      description: update an existing Favourite Supplier
      operationId: updateFavouriteSupplier
      requestBody:
        description: favourite supplier to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavouriteSupplier'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouriteSupplier'
        '400':
          description: Invalid status value
  /api/favouriteSuppliers/request:
    post:
      tags:
        - favouriteSupplier
      summary: send registration request to a supplier
      description: send registration request to a supplier
      operationId: sendRegistrationRequest
      requestBody:
        description: registration request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavouriteSupplierRegistrationRequest'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /api/favouriteSuppliers/{supplierId}/request:
    post:
      tags:
        - favouriteSupplier
      summary: send registration request to a supplier
      description: send registration request to a supplier
      operationId: sendRegistrationRequestWithoutCompanyId
      parameters:
        - in: path
          name: supplierId
          schema:
            type: number
          required: true
          description: identifier of the Favourite Product to be deleted

      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /api/favouriteProducts:
    post:
      tags:
        - favouriteProduct
      summary: create a new Favourite Product
      description: create a new Favourite Product
      operationId: createFavouriteProduct
      requestBody:
        description: new favourite product to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavouriteProduct'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouriteProduct'
        '400':
          description: Invalid status value
    put:
      tags:
        - favouriteProduct
      summary: update an existing Favourite Product
      description: update an existing Favourite Product
      operationId: updateFavouriteProduct
      requestBody:
        description: favourite product to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavouriteProduct'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouriteProduct'
        '400':
          description: Invalid status value
  /api/favouriteProducts/{favouriteProductId}:
    delete:
      tags:
        - favouriteProduct
      summary: delete an existing Favourite Product
      description: delete an existing Favourite Product
      operationId: deleteFavouriteProduct
      parameters:
        - in: path
          name: favouriteProductId
          schema:
            type: number
          required: true
          description: identifier of the Favourite Product to be deleted
      responses:
        '204':
          description: successful operation
        '400':
          description: Invalid status value
  /api/suppliers:
    get:
      tags:
        - supplier
      summary: Get all suppliers
      description: Get all suppliers without products
      operationId: getSuppliers
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                description: suppliers
                items:
                  $ref: '#/components/schemas/Supplier'
        '400':
          description: Invalid status value
  /api/suppliers/{supplierId}/products:
    get:
      tags:
        - supplier
      summary: get all products for a given supplier
      description: get all products for a given supplier
      operationId: getSupplierProductsBySupplierId
      parameters:
        - in: path
          name: supplierId
          schema:
            type: number
          required: true
          description: identifier of the Supplier
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplierProduct'
        '400':
          description: Invalid status value
...
  /api/public/supplierOrderConfirmation/emailApproval/{encodedSupplierOrderId}:
    put:
      tags:
        - public
      summary: confirm an Order
      description: this is for supplier to confirm an Order
      operationId: confirmOrderFromEmail
      parameters:
        - in: path
          name: encodedSupplierOrderId
          schema:
            type: string
          required: true
          description: base64 encoded Identifier of the Order to be confirmed
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /api/public/register:
    post:
      tags:
        - public
      summary: register a company and a company owner
      description: this is for user to register their business
      operationId: registerCompany
      requestBody:
        description: company user to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Registration'

      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /api/public/supplierOrderConfirmation/emailRejection/{encodedSupplierOrderId}:
    put:
      tags:
        - public
      summary: reject an Order from email by a supplier
      description: this is for email supplier to reject an Order
      operationId: rejectOrderFromEmailSupplier
      parameters:
        - in: path
          name: encodedSupplierOrderId
          schema:
            type: string
          required: true
          description: base64 encoded Identifier of the Order to be rejected
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /api/public/supplierOrderConfirmation/externalApproval:
    put:
      tags:
        - public
      summary: confirm an Order from an integrated supplier
      description: this is for integrated supplier to confirm an Order
      operationId: confirmOrderFromIntegratedSupplier
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: identifier of the Order to be confirmed
        - in: query
          name: token
          schema:
            type: string
          required: true
          description: secure token
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /api/public/supplierOrderConfirmation/externalRejection:
    put:
      tags:
        - public
      summary: reject an Order from an integrated supplier
      description: this is for integrated supplier to reject an Order
      operationId: rejectOrderFromIntegratedSupplier
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
          description: identifier of the Order to be rejected
        - in: query
          name: token
          schema:
            type: string
          required: true
          description: secure token
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
  /api/public/supplierOrderConfirmation/externalApprovalWithUpdate:
    put:
      tags:
        - public
      summary: update and confirm an Order from an integrated supplier
      description: this is for integrated supplier to update and confirm an Order
      operationId: updateAndConfirmOrderFromIntegratedSupplier
      parameters:
        - in: query
          name: token
          schema:
            type: string
          required: true
          description: secure token
      requestBody:
        description: updated order
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CAFSupplierOrder'
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid status value
...
components:
  schemas:
...
    Supplier:
      type: object
      properties:
        id:
          type: number
          description: identifier of the Supplier
        supplierName:
          type: string
          description: supplier name
        description:
          type: string
          description: supplier description
        cutOff:
          type: string
          description: date or time that the supplier stop taking new orders
        supplierLogoUrl:
          type: string
          description: logo url for the Supplier
        supplierEmail:
          type: string
          description: supplier email
        contactNumber:
          type: string
          description: supplier contact number
        addressLine1:
          type: string
          description: supplier address line1
        addressLine2:
          type: string
          description: supplier address line2
        addressLine3:
          type: string
          description: supplier address line3
        deliveryDay:
          type: string
          description: on what days the supplier make delivery
        minOrder:
          type: string
          description: min order amount
        category:
          type: string
          description: supplier category
        shopName:
          type: string
          description: supplier shopName
        accountNumber:
          type: string
          description: supplier accountNumber
        isIntegratedSupplier:
          type: boolean
          description: whether supplier is an integrated supplier or not
        supplierProducts:
          type: array
          description: products the supplier sell
          items:
            $ref: '#/components/schemas/SupplierProduct'
    SupplierProduct:
      type: object
      properties:
        id:
          type: number
          description: identifier of the Product
        supplierId:
          type: number
          description: identifier of the Product's Supplier
        productName:
          type: string
          description: product name
        description:
          type: string
          description: product description
        productImageUrl:
          type: string
          description: image url for the product
        nickname:
          type: string
          description: nickname of the product
        unit:
          type: string
          description: unit of the product
        vat:
          type: number
          description: VAT of the Product
        price:
          type: number
          description: unit price of the Product
...
    FavouriteSupplier:
      required:
        - supplierId
        - shopName
      type: object
      properties:
        id:
          type: number
          description: identifier of the Favourite Supplier
        supplierId:
          type: number
          description: identifier of the Supplier linked to this Favourite
        shopName:
          type: string
          description: name of the current Company in the Supplier
        accountNumber:
          type: string
          description: account number of the current Company in the Supplier
    FavouriteSupplierRegistrationRequest:
      required:
        - supplierId
        - companyName
        - companyEmail
        - address
      type: object
      properties:
        supplierId:
          type: number
          description: identifier of the supplier
        companyName:
          type: string
          description: name of the Company
        companyEmail:
          type: string
          description: Company contact email
        address:
          type: string
          description: Company address
    FavouriteProduct:
      required:
        - supplierId
        - supplierProductId
        - nickname
      type: object
      properties:
        id:
          type: number
          description: identifier of the Favourite Product
        supplierId:
          type: number
          description: identifier of the Supplier linked to this Favourite Product
        supplierProductId:
          type: number
          description: identifier of the Supplier Product linked to this Favourite Product
        nickname:
          type: string
          description: nickname / display name of the Favourite
...
    CAFSupplierOrder:
      type: object
      required:
        - id
        - deliveryDate
      properties:
        id:
          type: number
          description: identifier of the Order
        deliveryDate:
          type: string
          description: expected delivery date for the order, it is in ISO-8601 format, e.g. 2007-12-03
        totalVat:
          type: number
          description: total VAT of this Order
        totalPrice:
          type: number
          description: total price of this Order
        supplierorderItems:
          type: array
          description: items in this Order
          items:
            $ref: '#/components/schemas/CAFSupplierOrderItem'
    CAFSupplierOrderItem:
      type: object
      required:
        - supplierProductId
        - supplierProductName
        - quantity
        - vat
        - price
      properties:
        supplierProductId:
          type: number
          description: identifier of the linked SupplierProduct
        supplierProductName:
          type: string
          description: name of the linked SupplierProduct
        quantity:
          type: number
          description: quantity of the OrderItem
        vat:
          type: number
          description: VAT of the OrderItem
        price:
          type: number
          description: unit price of the OrderItem
    Registration:
      type: object
      required:
        - company
        - companyUser
      properties:
        company:
          type: object
          description: parameters of the Company
          $ref: '#/components/schemas/CompanyRegister'
        companyUser:
          type: object
          description: parameters of the CompanyUser
          $ref: '#/components/schemas/CompanyUserRegister'
    CompanyRegister:
      type: object
      required:
        - companyName
        - companyEmail
        - addressLine1
        - contactName
        - contactNumber

      properties:
        id:
          type: number
          description: identifier of the Company
        companyName:
          type: string
          description: name of the Company
        companyEmail:
          type: string
          description: email of the Company
        contactName:
          type: string
          description: contact name of the Company
        contactNumber:
          type: string
          description: contact number of the Company
        addressLine1:
          type: string
          description: Company address line1
        addressLine2:
          type: string
          description: Company address line2
        addressLine3:
          type: string
          description: Company address line3
    CompanyUserRegister:
      required:
        - firstName
        - lastName
        - email
      type: object
      properties:
        id:
          type: number
          description: identifier of the team member
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        contactNumber:
          type: string
...